+++ SEN4992/Pages/AddView.swift
@@ -1,0 +1,28 @@
+//
+//  AddView.swift
+//  SEN4992
+//
+//  Created by Oğuz Kaya on 20.04.2022.
+//
+
+import SwiftUI
+
+struct AddView: View {
+    //MARK: - PROPERTIES
+    
+    
+    
+    
+    
+    
+    //MARK: - BODY
+    var body: some View {
+        Text("Hello, World!\nAddView")
+    }
+}
+
+struct AddView_Previews: PreviewProvider {
+    static var previews: some View {
+        AddView()
+    }
+}
