+++ SEN4992/Pages/MainView.swift
@@ -8,25 +8,37 @@
 import SwiftUI
 
 struct MainView: View {
-    let userApi = UserApi()
+    //MARK: - PROPERTIES
     
+    
+    
+    init() {
+    }
+    //MARK: - BODY
     var body: some View {
-                VStack {
-                    Text("Hello, World!\nuser id:  \(userApi.currentUserId)\nname: \(userApi.currentUserName)")
-                    CustomButtonUI(function: {
-                        userApi.logOut {
-                            userApi.userLoginPageCheck()
-                        } onError: { err in
-                            print(err)
-                        }
-        
-                    }, title: "LogOut")
+        TabView {
+                TodayView()
+                    .tabItem {
+                        Image(systemName: "house")
+                        Text("Today")
                 }
+                AddView()
+                    .tabItem {
+                        Image(systemName: "plus.circle")
+                        Text("Add")
+                }
+                HistoryView()
+                    .tabItem {
+                        Image(systemName: "list.bullet.rectangle")
+                        Text("History")
+                }
+            }//: tabview
     }
 }
 
 struct MainView_Previews: PreviewProvider {
     static var previews: some View {
         MainView()
+            .preferredColorScheme(.light)
     }
 }
