+++ SEN4992/Pages/TodayView.swift
@@ -1,0 +1,29 @@
+//
+//  TodayView.swift
+//  SEN4992
+//
+//  Created by Oğuz Kaya on 20.04.2022.
+//
+
+import SwiftUI
+
+struct TodayView: View {
+    //MARK: - PROPERTIES
+    
+    
+    
+    
+    
+    
+    
+    //MARK: - BODY
+    var body: some View {
+        Text("Hello, World!\nToday View")
+    }
+}
+
+struct TodayView_Previews: PreviewProvider {
+    static var previews: some View {
+        TodayView()
+    }
+}
